Add files via upload
The sonars sensors attached to the front bottom of the car measure the distance in cm between the surface of the road and the sensors and send the data to the ATmega-328. The ATmega-328 is coded to store the incoming value in two float type variables named sonar1 and sonar2. The two variables are passed to the ThingSpeak server with the help of ESP-8266 which is three mode Wi-Fi module.   Specifications:  Following are the components used in the project:  • ATmega-328. • Crystal Oscillator (16,000 MHz).  • Arduino UNO. • Sonars (x2). • Resistors.  • Capacitors. • ESP 8266 Wi-Fi module. • 4WD car chassis kit.  • L298 motor drivers (x2). • DC motors (x4). 
• Lithium polymer battery. • Switches. • LEDs. • Buzzer. • Jumper wires. • Programming environment = Arduino IDE.